feat: implementado regra para movimentacao de pecas no estoque
Adicionado metodo para Criar ou Editar uma ou mais pecas em um orcamento;

Removido metodo UpdateOrcamento para simplificar as coisas;

Agora a API lida com movimentacoes no estoque. Ao criar um orcamento, a API vai checar o estoque para poder  adicionar a peca no orcamento do cliente; A API tambem ira adicionar ou remover a quantidade no estoque, juntamente com o estado da peca (se esta em estoque, reservada ou sem estoque.)

Atualizado metodo RemoveOrcamento para lidar com essas mudancas;

+++ Ultracar/Repository/OrcamentoRepository.cs
@@ -15,10 +15,85 @@ namespace Ultracar.Repository
 
     //
 
+    public void AddOrUpdatePartInOrcamento(int orcamentoId, Peca peca)
+    {
+      Orcamento? orcamento = _context.Orcamentos
+      .Include(obj => obj.Pecas!)
+      .ThenInclude(obj => obj.Estoque)
+      .FirstOrDefault(orcament => orcament.Id == orcamentoId);
+
+      if (orcamento == null)
+      {
+        throw new InvalidOperationException("[Ultracar] - ERROR: failed to quote. (are you specifying the quote 'id' in the request body?)");
+      }
+
+      if (orcamento.Pecas == null)
+      {
+        throw new InvalidOperationException("[Ultracar] - ERROR: failed to find, 'Peca' table is empty or null.");
+      }
+
+      // check if part to be added or modified already exists in the quote
+      Peca? foundPart = orcamento.Pecas.FirstOrDefault(part => part.Id == peca.Id);
+
+      // if it doesn't exist
+      if (foundPart == null)
+      {
+        // add peca in the client's quote
+        orcamento.Pecas.Add(peca);
+
+        // save changes in the data base
+        _context.SaveChanges();
+      }
+      else // if it does exist
+      { 
+        // update data of existing part 
+        foundPart.EstoqueId = peca.EstoqueId;
+        foundPart.OrcamentoId = peca.OrcamentoId;
+        foundPart.Quantidade = peca.Quantidade;
+        foundPart.NomePeca = peca.NomePeca;
+      }
+
+      // check if part to be added or modified exists in stock
+      // if the part is valid, the API now has a valid instance of Estoque
+      Estoque? foundPartInStock = _context.Estoque.FirstOrDefault(stock => stock.Id == peca.EstoqueId);
+
+      // in case the stock is invalid
+      if (foundPartInStock == null)
+      {
+        throw new InvalidOperationException($"[Ultracar] - ERROR: failed to validade stock, part with id: {peca.EstoqueId} was not found in stock.");
+      }
+      else
+      {
+        // NOTE: this below is a very subject rule, since in theory the establishment could make a quote, add the parts to it
+        // but the client might not have authorized the order so the part could still be in stock
+
+        // check if the part actually has stock available
+        if (foundPartInStock.EstoquePeca >= peca.Quantidade)
+        {
+          // set part state to Reserve since it's being added into a clients quote
+          foundPartInStock.TipoMovimentacao = ActionTypes.Reserved;
+
+          // deduct the quantity of part in the stock
+          foundPartInStock.EstoquePeca -= peca.Quantidade;
+
+        }
+        else
+        {
+          // set part status to out of stock
+          foundPartInStock.TipoMovimentacao = ActionTypes.OutOfStock;
+
+          throw new InvalidOperationException($"[Ultracar] - ERROR: failed to add part, the {foundPartInStock.NomePeca} stock is empty or it doesn't have the amount specified in the request.");
+        }
+      }
+    }
+
+    //
+
     public IEnumerable<OrcamentoDto> GetOrcamentos()
     {
       List<OrcamentoDto> orcamentos = _context.Orcamentos
-        .Include(obj => obj.Pecas)
+        .Include(obj => obj.Pecas!)
+        .ThenInclude(obj => obj.Estoque)
         .Select(orcament => new OrcamentoDto
         {
             Id = orcament.Id,
@@ -28,7 +103,8 @@ namespace Ultracar.Repository
             Pecas = orcament.Pecas!.Select(peca => new PecaDto
             {
                 Id = peca.Id,
-                NomePeca = peca.NomePeca,
+                EstoqueId = peca.EstoqueId,
+                NomePeca = peca.Estoque!.NomePeca,
                 Quantidade = peca.Quantidade,
                 PecaEntregue = peca.PecaEntregue
             }).ToList()
@@ -44,7 +120,8 @@ namespace Ultracar.Repository
     public OrcamentoDto GetOrcamentoById(int id)
     { 
       Orcamento? orcamento = _context.Orcamentos
-      .Include(obj => obj.Pecas)
+      .Include(obj => obj.Pecas!)
+      .ThenInclude(obj => obj.Estoque)
       .FirstOrDefault(orcament => orcament.Id == id);
 
       if (orcamento == null) 
@@ -61,6 +138,7 @@ namespace Ultracar.Repository
         Pecas = orcamento.Pecas?.Select(peca => new PecaDto
         {
           Id = peca.Id,
+          EstoqueId = peca.EstoqueId,
           NomePeca = peca.NomePeca,
           Quantidade = peca.Quantidade,
           PecaEntregue = peca.PecaEntregue
@@ -72,7 +150,8 @@ namespace Ultracar.Repository
     public IEnumerable<OrcamentoDto> GetOrcamentoByName(string clientName)
     {
       List<OrcamentoDto> orcamentos = _context.Orcamentos
-        .Include(obj => obj.Pecas)
+        .Include(obj => obj.Pecas!)
+        .ThenInclude(obj => obj.Estoque)
         .Where(orc => orc.NomeCliente == clientName)
         .Select(orcamento => new OrcamentoDto
         {
@@ -83,7 +162,8 @@ namespace Ultracar.Repository
           Pecas = orcamento.Pecas!.Select(peca => new PecaDto
           {
             Id = peca.Id,
-            NomePeca = peca.NomePeca,
+            EstoqueId = peca.EstoqueId,
+            NomePeca = peca.Estoque!.NomePeca,
             Quantidade = peca.Quantidade,
             PecaEntregue = peca.PecaEntregue
           }).ToList()
@@ -99,7 +179,8 @@ namespace Ultracar.Repository
     public IEnumerable<OrcamentoDto> GetOrcamentoByLicensePlate(string licensePlate)
     {
       List<OrcamentoDto> orcamentos = _context.Orcamentos
-        .Include(obj => obj.Pecas)
+        .Include(obj => obj.Pecas!)
+        .ThenInclude(obj => obj.Estoque)
         .Where(orc => orc.PlacaVeiculo == licensePlate)
         .Select(orcamento => new OrcamentoDto
         {
@@ -110,7 +191,8 @@ namespace Ultracar.Repository
           Pecas = orcamento.Pecas!.Select(peca => new PecaDto
           {
             Id = peca.Id,
-            NomePeca = peca.NomePeca,
+            EstoqueId = peca.EstoqueId,
+            NomePeca = peca.Estoque!.NomePeca,
             Quantidade = peca.Quantidade,
             PecaEntregue = peca.PecaEntregue
           }).ToList()
@@ -126,7 +208,8 @@ namespace Ultracar.Repository
     public OrcamentoDto GetOrcamentoByNumber(string orcamentoNumber)
     { 
       Orcamento? orcamento = _context.Orcamentos
-      .Include(obj => obj.Pecas)
+      .Include(obj => obj.Pecas!)
+      .ThenInclude(obj => obj.Estoque)
       .FirstOrDefault(orcament => orcament.NumeracaoOrcamento == orcamentoNumber);
 
       if (orcamento == null) 
@@ -143,6 +226,7 @@ namespace Ultracar.Repository
         Pecas = orcamento.Pecas?.Select(peca => new PecaDto
         {
           Id = peca.Id,
+          EstoqueId = peca.EstoqueId,
           NomePeca = peca.NomePeca,
           Quantidade = peca.Quantidade,
           PecaEntregue = peca.PecaEntregue
@@ -156,34 +240,41 @@ namespace Ultracar.Repository
 
     public OrcamentoDto UpdateOrcamentoById(int id, Orcamento orcamentoBody)
     {   
-      if (orcamentoBody == null)
+     if (orcamentoBody == null)
       {
         throw new InvalidOperationException("[Ultracar] - ERROR: failed to update, body is empty.");
       }
 
       // find quote to edit by its id
       Orcamento? updatedOrcamento = _context.Orcamentos
-      .Include(obj => obj.Pecas)
+      .Include(obj => obj.Pecas!)
+      .ThenInclude(obj => obj.Estoque)
       .FirstOrDefault(orcament => orcament.Id == id);
 
       if (updatedOrcamento == null) 
       {
-        throw new InvalidOperationException("[Ultracar] - ERROR: failed to update, quote(s) not found.");
+        throw new InvalidOperationException("[Ultracar] - ERROR: failed to update, quote not found.");
       }
 
-      // edit quote
-      _context.Orcamentos.Update(updatedOrcamento);
-
-      // edit quote with data received from body
+      // partially edit quote with data received from body
       updatedOrcamento.NumeracaoOrcamento = orcamentoBody.NumeracaoOrcamento;
       updatedOrcamento.NomeCliente = orcamentoBody.NomeCliente;
       updatedOrcamento.PlacaVeiculo = orcamentoBody.PlacaVeiculo;
 
+      if (orcamentoBody.Pecas != null)
+      {
+        // iterate through all the parts being added in the quote body
+        foreach (Peca? peca in orcamentoBody.Pecas)
+        {
+          // update part in the quote
+          AddOrUpdatePartInOrcamento(updatedOrcamento.Id, peca);
+        }
+      }
+
       // save changes in the data base
       _context.SaveChanges();
       
-      // create a simple dto to display the changes
-      // partialy at the moment
+      // create dto to display the changes
       OrcamentoDto result = new()
       {
         Id = updatedOrcamento.Id,
@@ -193,38 +284,7 @@ namespace Ultracar.Repository
         Pecas = updatedOrcamento.Pecas?.Select(peca => new PecaDto
         {
           Id = peca.Id,
-          NomePeca = peca.NomePeca,
-          Quantidade = peca.Quantidade,
-          PecaEntregue = peca.PecaEntregue
-        }).ToList(),
-      };
-
-      return result;
-    }
-    public OrcamentoDto UpdateOrcamento(Orcamento orcamentoBody)
-    { 
-      if (orcamentoBody == null) 
-      {
-        throw new InvalidOperationException("[Ultracar] - ERROR: failed to update, body is empty.");
-      }
-
-      // edit an entire quote column
-      _context.Orcamentos.Update(orcamentoBody);
-
-      // save changes in the data base
-      _context.SaveChanges();
-
-      // create a simple dto to display the changes
-      // partialy at the moment
-      OrcamentoDto result = new()
-      {
-        Id = orcamentoBody.Id,
-        NumeracaoOrcamento = orcamentoBody.NumeracaoOrcamento,
-        NomeCliente = orcamentoBody.NomeCliente,
-        PlacaVeiculo = orcamentoBody.PlacaVeiculo,
-        Pecas = orcamentoBody.Pecas?.Select(peca => new PecaDto
-        {
-          Id = peca.Id,
+          EstoqueId = peca.EstoqueId,
           NomePeca = peca.NomePeca,
           Quantidade = peca.Quantidade,
           PecaEntregue = peca.PecaEntregue
@@ -243,13 +303,41 @@ namespace Ultracar.Repository
         throw new InvalidOperationException("[Ultracar] - ERROR: failed to create, body is empty.");
       }
 
-      // populate quote table with body content
+      if (newOrcamentoBody.Pecas == null)
+      {
+        throw new InvalidOperationException("[Ultracar] - ERROR: failed to access the part(s) in quote, 'Pecas' table is empty or null.");
+      }
+
+      // populate dummy list to store current parts in the quote
+      // before the api deletes them to handle the request
+      List<Peca> pecasToAdd = newOrcamentoBody.Pecas;
+      
+      if (newOrcamentoBody.Pecas != null)
+      {
+        // remove parts from new quote body to avoid conflict
+        newOrcamentoBody.Pecas = null;
+      }
+
+      // partialy populate current quote column with body content
       _context.Orcamentos.Add(newOrcamentoBody);
 
       // save changes in the data base
       _context.SaveChanges();
 
-      // create a simple dto to display the created quote
+      if (pecasToAdd != null)
+      {
+        // iterate through all the parts being added in the quote body
+        foreach (Peca? peca in pecasToAdd)
+        {
+          // add new part(s) in the quote
+          AddOrUpdatePartInOrcamento(newOrcamentoBody.Id, peca);
+        }
+      }
+
+      // save changes in the data base
+      _context.SaveChanges();
+
+      // create dto to display the created quote
       OrcamentoDto result = new()
       {
         Id = newOrcamentoBody.Id,
@@ -259,6 +347,7 @@ namespace Ultracar.Repository
         Pecas = newOrcamentoBody.Pecas?.Select(peca => new PecaDto
         {
           Id = peca.Id,
+          EstoqueId = peca.EstoqueId,
           NomePeca = peca.NomePeca,
           Quantidade = peca.Quantidade,
           PecaEntregue = peca.PecaEntregue
@@ -274,7 +363,8 @@ namespace Ultracar.Repository
     {
       // find quote to remove by its id
       Orcamento? orcamentoToRemove = _context.Orcamentos
-      .Include(obj => obj.Pecas)
+      .Include(obj => obj.Pecas!)
+      .ThenInclude(obj => obj.Estoque)
       .FirstOrDefault(orcament => orcament.Id == id);
 
       if (orcamentoToRemove == null) 
@@ -282,7 +372,28 @@ namespace Ultracar.Repository
         throw new InvalidOperationException("[Ultracar] - ERROR: failed to remove, quote not found.");
       }
 
-      // remove orcament 
+      if (orcamentoToRemove.Pecas == null)
+      {
+        throw new InvalidOperationException("[Ultracar] - ERROR: failed to access the part(s) in quote, 'Pecas' table is empty or null.");
+      }
+
+      foreach (Peca? peca in orcamentoToRemove.Pecas)
+      {
+        // find part by its id in stock
+        Estoque? foundPartInStock = _context.Estoque.FirstOrDefault(stock => stock.Id == peca.EstoqueId);
+
+        if (foundPartInStock != null)
+        {
+          // add the quantity of parts being removed back to the stock
+          foundPartInStock.EstoquePeca += peca.Quantidade;
+
+          // set part state to InStock since it's being returned or 
+          // if the part was out of stock because it got placed in a client's quote
+          foundPartInStock.TipoMovimentacao = ActionTypes.InStock;
+        }
+      }
+
+      // remove quote
       _context.Orcamentos.Remove(orcamentoToRemove);
 
       // save changes in the data base
